chore: refresh user datagrid on action
