+++ client/app/components/datagrid/cells/ChangeUserOperatorStatusColumnCell.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useState } from "react";
 import { GridRenderCellParams } from "@mui/x-data-grid/models/params/gridCellParams";
 import Button, { ButtonOwnProps } from "@mui/material/Button";
 import { actionHandler } from "@/app/utils/actions";
@@ -54,7 +53,6 @@ const handleUpdateStatus = async (
 const ChangeUserOperatorStatusColumnCell = (params: ButtonRenderCellParams) => {
   const userOperatorStatus = params.row.status;
   const userOperatorId = params.row.userOperatorId;
-  const [statusState, setStatusState] = useState(userOperatorStatus);
 
   const buttonsToShow = (status: Status): UserOperatorStatusAction[] => {
     if (status === Status.MYSELF) {
@@ -88,13 +86,22 @@ const ChangeUserOperatorStatusColumnCell = (params: ButtonRenderCellParams) => {
 
   return (
     <Stack direction="row" spacing={1}>
-      {buttonsToShow(statusState).map((item, index) => (
+      {buttonsToShow(userOperatorStatus).map((item, index) => (
         <Button
           variant={item.title === "Undo" ? "text" : "outlined"}
           key={index}
           onClick={async () => {
             const res = await handleUpdateStatus(userOperatorId, item.statusTo);
-            setStatusState(res.status);
+            params.api.updateRows([
+              {
+                ...params.row,
+                accessType:
+                  res.status === Status.DECLINED
+                    ? "N/A"
+                    : params.row.accessType,
+                status: res.status,
+              },
+            ]);
           }}
           color={item.color}
           endIcon={item.icon}
